Setup skeleton module for all packages that mirror the dotnet project.
