--- libraries/bot-builder/src/main/resources/log4j2.json
@@ -1,25 +1,0 @@
-{
-  "configuration": {
-    "name": "Default",
-    "appenders": {
-      "Console": {
-        "name": "Console-Appender",
-        "target": "SYSTEM_OUT",
-        "PatternLayout": {"pattern": "[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{1} - %msg%n"}
-      }
-    },
-    "loggers": {
-      "logger": [
-        {
-          "name": "com.microsoft.bot.builder",
-          "level": "debug",
-          "appender-ref": [{"ref": "Console-Appender", "level": "debug"}]
-        }
-      ],
-      "root": {
-        "level": "warn",
-        "appender-ref": {"ref": "Console-Appender","level": "warn"}
-      }
-    }
-  }
-}
