+++ libraries/bot-configuration/pom.xml
@@ -10,12 +10,12 @@
     <relativePath>../../pom.xml</relativePath>
   </parent>
 
-  <artifactId>bot-builder</artifactId>
+  <artifactId>bot-configuration</artifactId>
   <packaging>jar</packaging>
   <version>4.0.0-SNAPSHOT</version>
 
   <name>${project.groupId}:${project.artifactId}</name>
-  <description>Bot Framework Connector</description>
+  <description>Bot Framework Dialogs</description>
   <url>https://dev.botframework.com/</url>
 
   <licenses>
@@ -54,48 +54,6 @@
     <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.microsoft.rest</groupId>
-      <artifactId>client-runtime</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.microsoft.azure</groupId>
-      <artifactId>azure-client-runtime</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.microsoft.azure</groupId>
-      <artifactId>azure-client-authentication</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.module</groupId>
-      <artifactId>jackson-module-parameter-names</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jdk8</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jsr310</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.auth0</groupId>
-      <artifactId>java-jwt</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.auth0</groupId>
-      <artifactId>jwks-rsa</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.microsoft.bot</groupId>
-      <artifactId>bot-schema</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.microsoft.bot</groupId>
-      <artifactId>bot-connector</artifactId>
     </dependency>
   </dependencies>
 
@@ -151,7 +109,7 @@
             <groupId>org.codehaus.mojo</groupId>
             <artifactId>cobertura-maven-plugin</artifactId>
             <configuration>
-              <outputDirectory>../../cobertura-report/bot-builder</outputDirectory>
+              <outputDirectory>../../cobertura-report/bot-configuration</outputDirectory>
               <format>xml</format>
               <maxmem>256m</maxmem>
               <!-- aggregated reports for multi-module projects -->
