+++ libraries/bot-dialogs/src/main/java/com/microsoft/bot/dialogs/prompts/Choice.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.microsoft.bot.builder.prompts;
+package com.microsoft.bot.dialogs.prompts;
 
 import com.microsoft.bot.schema.CardAction;
