+++ libraries/bot-integration-core/pom.xml
@@ -57,16 +57,10 @@
     <dependency>
       <groupId>com.microsoft.bot</groupId>
       <artifactId>bot-schema</artifactId>
-      <version>4.0.0-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>com.microsoft.bot</groupId>
       <artifactId>bot-connector</artifactId>
-      <version>4.0.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
     </dependency>
   </dependencies>
